# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GST Billing & Inventory Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      summary: Update invoice payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # Purchases
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Dashboard
...
security:
  # Every operation requires a session unless it overrides this with
  # `security: []`. The spec previously declared no authentication at all, so
  # the generated clients — and anyone reading it — saw a wholly public API.
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gst_session
      description: >-
        Session cookie issued by /auth/login and /auth/register. HttpOnly, so
        page scripts cannot read it. Browser clients must also send the CSRF
        token from the readable `gst_csrf` cookie in an `X-CSRF-Token` header
        on every state-changing request.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        For native clients that have no cookie jar. Browser clients should use
        the cookie instead, so the token is never reachable from script.
  schemas:
...
    InvoiceItem:
      type: object
      required: [quantity, gstRate]
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        description:
          type: string
          description: Line description. This is what clients send.
        productName:
          type: string
          deprecated: true
          description: Legacy alias for `description`. Accepted, never returned.
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        unitPrice:
          type: number
          description: Price per unit. This is what clients send.
        rate:
          type: number
          deprecated: true
          description: Legacy alias for `unitPrice`. Accepted, never returned.
        discount:
          type: number
        gstRate:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        taxableAmount:
          type: number
        totalAmount:
          type: number
    Invoice:
      type: object
      required: [id, invoiceNumber, type, status, customerId, customerName, invoiceDate, subtotal, totalGst, grandTotal]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        type:
          type: string
        status:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        customerGstin:
          type: ["string", "null"]
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        placeOfSupply:
          type: ["string", "null"]
        isInterstate:
          type: boolean
        subtotal:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalGst:
          type: number
        roundOff:
          type: number
        grandTotal:
          type: number
        paidAmount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
    InvoiceListResponse:
      type: object
      required: [invoices, total]
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
    InvoiceInput:
      type: object
      required: [invoiceDate, items]
      properties:
        type:
          type: string
        customerId:
          type: integer
          description: >-
            Omit for a walk-in customer and send `customerName` instead. The id
            must belong to the caller's own business.
        customerName:
          type: string
          description: Required when `customerId` is omitted.
        customerGstin:
          type: string
          nullable: true
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        placeOfSupply:
          type: ["string", "null"]
        isInterstate:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceUpdate:
      type: object
      properties:
        type:
          type: string
        customerId:
          type: integer
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        placeOfSupply:
          type: ["string", "null"]
        isInterstate:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [paid, unpaid, partial, cancelled]
        paymentStatus:
          type: string
          enum: [paid, unpaid, partial, cancelled]
          description: Alias for `status`; whichever is present is used.
        paidAmount:
          type: ["number", "null"]
...
    Payment:
      type: object
      required: [id, type, amount, date, mode]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        date:
          type: string
        mode:
          type: string
        referenceNumber:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        vendorId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentListResponse:
      type: object
      required: [payments, total]
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
    PaymentInput:
      type: object
      required: [type, amount, date, mode]
      properties:
        type:
          type: string
        amount:
          type: number
        date:
          type: string
        mode:
          type: string
        referenceNumber:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        vendorId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...