# lib/api-zod/src/generated/types/invoiceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GST Billing & Inventory Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceItem {
  id?: number;
  /** @nullable */
  productId?: number | null;
  /** Line description. This is what clients send. */
  description?: string;
  /**
     * Legacy alias for `description`. Accepted, never returned.
     * @deprecated
     */
  productName?: string;
  /** @nullable */
  hsnCode?: string | null;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** Price per unit. This is what clients send. */
  unitPrice?: number;
  /**
     * Legacy alias for `unitPrice`. Accepted, never returned.
     * @deprecated
     */
  rate?: number;
  discount?: number;
  gstRate: number;
  cgst?: number;
  sgst?: number;
  igst?: number;
  taxableAmount?: number;
  totalAmount?: number;
}