# lib/api-zod/src/generated/types/invoiceStatusInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GST Billing & Inventory Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceStatusInputPaymentStatus } from './invoiceStatusInputPaymentStatus';
import type { InvoiceStatusInputStatus } from './invoiceStatusInputStatus';

export interface InvoiceStatusInput {
  status?: InvoiceStatusInputStatus;
  /** Alias for `status`; whichever is present is used. */
  paymentStatus?: InvoiceStatusInputPaymentStatus;
  /** @nullable */
  paidAmount?: number | null;
}